refactor: reorganize auth DTOs and add post/comment features
- Moved auth-related request DTOs to dedicated auth package
- Added PostResponse DTO
- Implemented comment functionality with controller, service, and repository
- Added user management endpoints
- Added post management with media support
- Implemented follow and like functionality
- Added OpenAPI documentation
- Updated authentication service implementation

+++ backend/src/main/java/vn/pvhg/backend/response/ApiPaginatedResponse.java
@@ -11,12 +11,12 @@ public class ApiPaginatedResponse<T> {
     String message;
     boolean success;
     T data;
-    int page;
-    int pageSize;
-    int totalPages;
-    int totalElements;
+    long page;
+    long pageSize;
+    long totalPages;
+    long totalElements;
 
-    public ApiPaginatedResponse(HttpStatus status, String message, boolean success, T data, int page, int pageSize, int totalPages, int totalElements) {
+    public ApiPaginatedResponse(HttpStatus status, String message, boolean success, T data, long page, long pageSize, long totalPages, long totalElements) {
         this.status = status;
         this.message = message;
         this.success = success;
